[#1] Use Jsoup to parse the list of bus routes.
The list of bus routes is presented as an HTML <ul>, not as a JSON or
XML document, so we have to use an HTML parser to split it.

This also means that the route colouring information is not provided by
the server; the colours used by the website are arbitrary and hardcoded
in that UI. For consistency's sake, I will embed the same colour values
into this app, though that does pose a risk, in case the GRT changes its
colour scheme in the future.