[#1] Added a trip stop marker. Colored it to match the route colour.
I can use a ColorFilter to apply a colour transform to the base image,
so I don't need to have a unique image for each route.

Because the map use a BitmapDescription for the marker, and not a regular
drawable, I have to construct a Bitmap manually for the map to use. To
cut down on memory usage and CPU time, I keep a cache of the images used
for each route.